requirements.txt
anyio==4.12.1
attrs==21.2.0
Babel==2.8.0
beautifulsoup4==4.14.3
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.3.1
cloud-init==25.2
configobj==5.0.6
cryptography==3.4.8
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.1
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.5.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
PyGObject==3.42.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
regex==2026.1.15
requests==2.25.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.16.0
sos==4.9.2
soupsieve==2.8.3
ssh-import-id==5.11
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
triton==3.6.0
typer-slim==0.21.1
typing_extensions==4.15.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
